altered UserController & HomeController + tests
